fixed some issues

+++ ATM/Helpers/CurrencyConverter.cs
@@ -3,11 +3,11 @@
 namespace ATM_Machine;
 public static class CurrencyConverter
 {
-    private readonly static Dictionary<string, float> ConversionRates = new Dictionary<string, float>
+    private readonly static Dictionary<string, double> ConversionRates = new Dictionary<string, double>
     {
-        {"euro",2f },
-        {"usd",1.5f },
-        {"nat",1.25f }
+        {"euro",2 },
+        {"usd",1.5 },
+        {"nat",1.25 }
     };
     public static decimal Convert(string currency, decimal amount) => amount * (decimal)ConversionRates[currency];  
 }
